chocolate actions
