+++ src/App.js
@@ -2,22 +2,19 @@ import React from 'react';
 import '../src/style/style.css';
 import { Provider } from 'react-redux'
 import store from './redux/store'
-import CakeContainer from './components/CakeContainer';
 import IceCreamContainer from './components/IceCreamContainer';
-import NewCakeContainer from './components/NewCakeContainer'
-import ItemContainer from './components/ItemContainer';
+import CakeContainer from './components/CakeContainer'
 import UserContainer from './components/UserContainer';
 
 function App() {
   return (
     <Provider store = {store}>
       <div className="App">
-        <UserContainer />
-        <ItemContainer cake/>
-        <ItemContainer />
         <CakeContainer />
         <IceCreamContainer />
-        <NewCakeContainer />
+        <div style={{fontSize:20}}>
+          <UserContainer />
+        </div>
       </div>
     </Provider>
   );
