+++ src/components/CakeContainer.js
@@ -1,15 +1,17 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { buyCake } from "../redux"
 
 function CakeContainer(props) {
+    const [number, setNumber] = useState()
     return (
         <div>
             <h2>Number of cakes - {props.numOfCakes} </h2>
-            <button onClick= {props.buyCake}>Buy Cake</button>
+            <input type='text' value={number} onChange={e => setNumber(e.target.value)} />
+            <button onClick= {() => props.buyCake(number)}>Buy {number} Cake</button>
         </div>
     )
-}
+} 
 
 const mapStateToProps = state => {
     return{
@@ -19,30 +21,8 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
     return {
-        buyCake: () => dispatch(buyCake())
+        buyCake: (number) => dispatch(buyCake(number))
     }
 } 
 
 export default connect(mapStateToProps, mapDispatchToProps)(CakeContainer);
-
-//connect By Hook
-
-/*
-import React from 'react'
-import { useSelector, useDispatch } from "react-redux"
-import { buyCake } from '../redux'
-
-function HooksCakeContainer() {
-    const numOfCakes = useSelector(state => state.cake.numOfCakes)
-    const dispatch =  useDispatch()
-    return(
-        <div>
-        <h2>Num of cakes - {numOfCakes}</h2>
-        <button onClick={() => dispatch(buyCake())}>Buy Cake</button>
-        </div>
-    )
-}
-
-export default HooksCakeContainer
-
-*/
