+++ src/redux/cake/cakeAction.js
@@ -1,6 +1,6 @@
 import { BUY_CAKE } from './cakeTypes'
 
-export const buyCake = (number = 1) => {
+export const buyCake = (number) => {
     return {
         type: BUY_CAKE,
         payload: number
