+++ src/redux/chocolate/chocolateActions.js
@@ -1,0 +1,8 @@
+import { BUY_CHOCOLATE } from './cakeTypes'
+
+export const buyCHOCOLATE = (number) => {
+    return {
+        type: BUY_CHOCOLATE,
+        payload: number
+    }
+}
