+++ src/redux/chocolate/chocolateTypes.js
@@ -1,0 +1,1 @@
+export const BUY_CHOCOLATE = 'BUY_CHOCOLATE'
