+++ src/redux/iceCream/iceCreamAction.js
@@ -1,7 +1,8 @@
 import { BUY_ICECREAM } from './iceCreamTypes'
 
-export const buyIceCream = () => {
+export const buyIceCream = (number) => {
     return {
-        type: BUY_ICECREAM
+        type: BUY_ICECREAM,
+        payload: number
     }
 }
