+++ src/redux/iceCream/iceCreamReducer.js
@@ -8,7 +8,7 @@ const iceCreamReducer = (state = initialState, action) => {
     switch(action.type) {
         case BUY_ICECREAM: return {
             ...state,
-            numOfIceCreams: state.numOfIceCreams - 1
+            numOfIceCreams: state.numOfIceCreams - action.payload
         }
 
         default: return state
